requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver for migrations

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
email-validator==2.1.1

# Caching & Background Jobs
redis==5.0.1
celery==5.3.4

# Data Processing & ML
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
scipy==1.11.4  # Pinned to avoid resolution loop

# File Handling
python-multipart==0.0.6
pillow==10.1.0

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AWS Integration
boto3==1.34.22       # Pinned to match aiobotocore's botocore requirement
aiobotocore==2.11.0  # Async AWS SDK
aiofiles==23.2.1     # Async file operations

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0  # For async SQLite testing
httpx==0.25.2
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5
types-redis==4.6.0.20240106
types-python-jose==3.3.4.20240106

# Monitoring & Logging
structlog==23.2.0
psutil==5.9.6
rich==13.7.0  # For CLI tools